refactor setting conpile_context
